Move tools into lib/tools
