+++ lib/separators.js
@@ -1,0 +1,8 @@
+"use strict";
+
+module.exports = {
+    UP: "<", // one level up
+    SAME: ",", // same level
+    DOWN: ">", // one level down
+    RESET: "|", // reset level index and start new
+};
