+++ lib/serialize/parsePSList.js
