+++ lib/serialize/serializeTrie.js
@@ -1,11 +1,6 @@
 "use strict";
 
-const SEPARATORS = {
-    up: "<", // one level up
-    same: ",", // same level
-    down: ">", // one level down
-    reset: "|", // reset level index and start new
-};
+const SEPARATORS = require("../separators");
 
 function compareLinesAt(lineA, lineB, i) {
     const endOfLineA = i === lineA.length;
@@ -49,24 +44,24 @@ function lineToString(line, i, arr) {
         }
         if (indexOfDifference === 0) {
             // line and prevLine are completely different
-            separatorFromPrev = SEPARATORS.reset;
+            separatorFromPrev = SEPARATORS.RESET;
         } else if (prevLine.length === line.length && indexOfDifference === line.length - 1) {
             // only the last part of line and prevLine are different
-            separatorFromPrev = SEPARATORS.same;
+            separatorFromPrev = SEPARATORS.SAME;
         } else if (indexOfDifference > prevLine.length - 1) {
             // we don't need to go up the hierarchy first because prevLine is part of line
             // so let's just start with an initial down separator
-            separatorFromPrev = SEPARATORS.down;
+            separatorFromPrev = SEPARATORS.DOWN;
         } else {
             // line and prevLine are different, but share a common root at indexOfDifference - 1
             // we now need to go up the hierarchy to the common root
             separatorFromPrev = new Array(prevLine.length - indexOfDifference - 1)
-                .fill(SEPARATORS.up)
+                .fill(SEPARATORS.UP)
                 .join("");
         }
     }
 
-    return separatorFromPrev + line.slice(indexOfDifference).join(SEPARATORS.down);
+    return separatorFromPrev + line.slice(indexOfDifference).join(SEPARATORS.DOWN);
 }
 
 function serializeTrie(parsedList) {
@@ -88,7 +83,7 @@ function serializeTrie(parsedList) {
      *         /  \        /  \
      *       co   gov   静岡   岐阜
      *
-     * And the textual representation of the trie looks like (using SEPERATORS):
+     * And the textual representation of the trie looks like (using SEPARATORS):
      *
      * com,uk>co,gov<jp>静岡,岐阜<موقع
      */
